requirements.txt
# Build Tools (Install First)
setuptools>=65.0.0
wheel>=0.38.0

# Core ML Libraries
numpy>=1.26.0,<2.0.0
pandas>=1.5.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0

# MLflow & Tracking
mlflow>=2.8.1,<3.0.0
databricks-cli==0.18.0

# API & Deployment
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Data Processing
pyarrow>=15.0.0,<20.0.0
openpyxl>=3.1.0,<4.0.0

# Monitoring & Drift Detection
# Evidently updated to support Pydantic v2
evidently>=0.4.15,<0.5.0
prometheus-client==0.19.0
schedule==1.2.0
scipy>=1.11.0,<2.0.0

# Feature Store (Optional)
feast>=0.40.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
boto3==1.28.85

# Visualization
matplotlib==3.7.3
seaborn==0.12.2
plotly==5.17.0
